backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
psycopg2-binary>=2.9.0
google-cloud-storage>=1.40.0
pytest>=6.2.0
python-multipart>=0.0.5
requests>=2.26.0
pytz>=2021.1
python-dotenv>=1.0.0
APScheduler>=3.10.0
sendgrid>=6.11.0
PyJWT>=2.8.0
